server/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing
librosa>=0.10.0
numpy>=1.24.0
soundfile>=0.12.0

# Database
supabase>=2.0.0
pydantic>=2.0.0

# Vector Databases (optional - choose one or use mock)
qdrant-client>=1.7.0  # For Qdrant vector DB
pinecone-client>=2.2.0  # For Pinecone vector DB

# Environment and Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0  # For async testing

# Security
argon2-cffi>=23.1.0  # Argon2 password hashing
redis>=5.0.0  # Optional: distributed rate limiting
sqlalchemy>=2.0.0  # Database ORM for auth
asyncpg>=0.29.0  # Async PostgreSQL driver

# Blockchain
web3>=6.11.0  # Ethereum/Polygon interaction
eth-account>=0.11.0  # Account management

# Monitoring & Observability
prometheus-client>=0.19.0  # Prometheus metrics
opentelemetry-api>=1.21.0  # OpenTelemetry tracing
opentelemetry-sdk>=1.21.0  # OpenTelemetry SDK
opentelemetry-instrumentation-fastapi>=0.42b0  # FastAPI auto-instrumentation
opentelemetry-exporter-otlp-proto-grpc>=1.21.0  # OTLP gRPC exporter
gunicorn>=21.2.0  # Production WSGI server

# Optional: OpenAI for future embeddings
# openai>=1.0.0

# Development tools (optional)
# black>=23.11.0  # Code formatting
# flake8>=6.1.0  # Linting
# pytest-cov>=4.1.0  # Coverage reporting
